# back-end/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Cali Arena API
  description: |
    REST API for the Cali Arena tournament platform.

    Base URL: `http://localhost:8080`

    Everything under `/api` is served by the Spring Boot application (`app` module).

    All JSON responses are serialized from Kotlin data classes: dates use ISO-8601
    (e.g. `2026-08-20T10:00:00Z`), enums are serialized as strings, and `BigDecimal`
    fields are serialized as numbers.

    Error responses use [RFC 7807](https://tools.ietf.org/html/rfc7807) Problem
    Details with content type `application/problem+json`. Every error response has
    the shape:
    ```json
    {
      "type": "/problem/athlete-not-found",
      "title": "athlete-not-found",
      "status": 404
    }
    ```
  version: 1.0.0
  contact:
    name: Cali Arena
...
paths:
...
  /api/matches:
    post:
      tags: [Matches]
      summary: Create a match
      description: Requires the `ADMIN` role.
      operationId: createMatch
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchInput"
      responses:
        "201":
          description: Match created.
          headers:
            Location:
              description: URL of the created match.
              schema:
                type: string
                example: /api/matches/10
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
  /api/matches/{id}:
    get:
      tags: [Matches]
      summary: Get a match by id
      operationId: getMatchById
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: The match.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
  /api/matches/{id}/start:
    put:
      tags: [Matches]
      summary: Start a match
      description: Requires the `ADMIN` or `JUDGE` role. Also broadcasts a `STARTED` event on the judge websocket topic.
      operationId: startMatch
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: Initial progress after starting.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchProgress"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
        "409":
          $ref: "#/components/responses/ConflictProblem"
  /api/matches/{matchId}/reps:
    put:
      tags: [Matches]
      summary: Update red/blue reps for a running match
      description: Requires the `ADMIN` or `JUDGE` role.
      operationId: updateMatchReps
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/MatchIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRepsInput"
      responses:
        "202":
          description: Updated progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchProgress"
        "400":
          $ref: "#/components/responses/BadRequestProblem"
        "401":
          $ref: "#/components/responses/UnauthorizedProblem"
        "403":
          $ref: "#/components/responses/ForbiddenProblem"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
  /api/matches/judge/{judgeId}:
    get:
      tags: [Matches]
      summary: List all matches judged by a specific judge
      operationId: listMatchesByJudge
      parameters:
        - name: judgeId
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        "200":
          description: List of matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
  /api/matches/bracket/{id}:
    get:
      tags: [Matches]
      summary: List all matches of a bracket
      operationId: listMatchesByBracket
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: List of matches.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
        "404":
          $ref: "#/components/responses/NotFoundProblem"
  /api/matches/{id}/progress:
    get:
      tags: [Matches]
      summary: Get the current progress of a match
      operationId: getMatchProgress
      parameters:
        - $ref: "#/components/parameters/IdPath"
      responses:
        "200":
          description: The current progress.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchProgress"
        "404":
          $ref: "#/components/responses/NotFoundProblem"

  # ----------------------------------------------------------------- Brackets
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Token obtained from `POST /api/users/token`.
  parameters:
    IdPath:
      name: id
      in: path
      required: true
      description: Entity id.
      schema:
        type: integer
        format: int32
    MatchIdPath:
      name: matchId
      in: path
      required: true
      description: Match id.
      schema:
        type: integer
        format: int32
...
  responses:
    BadRequestProblem:
      description: Invalid request (e.g. invalid gender, insecure password, invalid role).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    UnauthorizedProblem:
      description: Missing or invalid Bearer token.
      headers:
        WWW-Authenticate:
          description: Authentication scheme expected by the server.
          schema:
            type: string
            example: bearer
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ForbiddenProblem:
      description: Authenticated but not allowed to perform the operation.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    NotFoundProblem:
      description: Referenced entity does not exist.
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    ConflictProblem:
      description: State conflict (e.g. duplicate name, invalid transition, mismatch).
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
  schemas:
    Problem:
      type: object
      description: RFC 7807 problem detail.
      required: [type, title, status]
      properties:
        type:
          type: string
          description: Stable identifier for the problem, e.g. `/problem/athlete-not-found`.
          example: /problem/athlete-not-found
        title:
          type: string
          description: Human-friendly title derived from the type.
          example: athlete-not-found
        status:
          type: integer
          format: int32
          description: HTTP status code.
          example: 404

    # ------------------------------------------------------------------ Inputs
...
    CreateMatchInput:
      type: object
      required: [bracketId, routineId, judgeId, athleteRedId, athleteBlueId]
      properties:
        bracketId:
          type: integer
          format: int32
        routineId:
          type: integer
          format: int32
        judgeId:
          type: integer
          format: int32
        athleteRedId:
          type: integer
          format: int32
        athleteBlueId:
          type: integer
          format: int32
    UpdateRepsInput:
      type: object
      description: Omit a side to keep its current value; `null` also keeps it.
      properties:
        redReps:
          type: integer
          format: int32
          nullable: true
          example: 15
        blueReps:
          type: integer
          format: int32
          nullable: true
          example: 12
...
    Match:
      type: object
      required: [id, bracketId, routineId, judgeId, status, createdAt]
      properties:
        id:
          type: integer
          format: int32
        bracketId:
          type: integer
          format: int32
        routineId:
          type: integer
          format: int32
        judgeId:
          type: integer
          format: int32
        athleteRedId:
          type: integer
          format: int32
          nullable: true
        athleteBlueId:
          type: integer
          format: int32
          nullable: true
        winnerAthleteId:
          type: integer
          format: int32
          nullable: true
        status:
          type: string
          enum: [PENDING, RUNNING, FINISHED]
        startedAt:
          type: string
          format: date-time
          nullable: true
        finishedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
    MatchProgress:
      type: object
      required: [id, matchId, redCurrentReps, blueCurrentReps, updatedAt]
      properties:
        id:
          type: integer
          format: int32
        matchId:
          type: integer
          format: int32
        redCurrentExerciseId:
          type: integer
          format: int32
          nullable: true
          description: "`null` once the red side has finished the routine."
        blueCurrentExerciseId:
          type: integer
          format: int32
          nullable: true
          description: "`null` once the blue side has finished the routine."
        redCurrentReps:
          type: integer
          format: int32
        blueCurrentReps:
          type: integer
          format: int32
        redFinishedAt:
          type: string
          format: date-time
          nullable: true
          description: Time the red side finished the routine (also the match end).
        blueFinishedAt:
          type: string
          format: date-time
          nullable: true
          description: Time the blue side finished the routine (also the match end).
        timerStartedAt:
          type: string
          format: date-time
          nullable: true
        timerRemainingSeconds:
          type: integer
          format: int32
          nullable: true
        updatedAt:
          type: string
          format: date-time
...